Split sync storage cleanup helper
